Computer Vision-Optical Character Recognition (OCR)/requirements.txt
azure.ai.vision.imageanalysis

pillow

azure-cognitiveservices-vision-computervision